feat: delete word

+++ src/controllers/entries.controller.ts
@@ -3,6 +3,7 @@ import { NextFunction, Response } from 'express';
 import entriesModel from '@models/entries.model';
 import { RequestWithUser } from '@/interfaces/auth.interface';
 import favoriteModel from '@/models/favorite.model';
+import { HttpException } from '@/exceptions/HttpException';
 
 class EntriesController {
   public entries = entriesModel;
@@ -25,16 +26,17 @@ class EntriesController {
   public saveFavoriteWord = async (req: RequestWithUser, res: Response, next: NextFunction) => {
     try {
       const word = String(req.params.word);
+      const userId = String(req.user._id);
 
-      const wordFound = await this.favorites.find({ userId: req.user._id.toString(), word });
+      const wordFound = await this.favorites.find({ userId, word });
 
       if (wordFound.length > 0) {
-        return res.status(200).json({ message: 'word is already in the favorites list' });
+        return res.status(200).json({ message: 'Word is already in the favorites list' });
       }
 
-      await this.favorites.create({ userId: req.user._id.toString(), word });
+      await this.favorites.create({ userId, word });
 
-      return res.status(200).json({ message: 'word was saved' });
+      return res.status(200).json({ message: 'Word was saved' });
     } catch (error) {
       next(error);
     }
@@ -43,9 +45,15 @@ class EntriesController {
   public removeFavoriteWord = async (req: RequestWithUser, res: Response, next: NextFunction) => {
     try {
       const word = String(req.params.word);
-      await this.favorites.deleteOne({ word });
+      const userId = String(req.user._id);
 
-      res.status(200).json({ message: 'word was deleted' });
+      const deletedWord = await this.favorites.deleteOne({ userId, word });
+
+      if (deletedWord.deletedCount === 0) {
+        throw new HttpException(404, 'Word not found');
+      }
+
+      res.status(200).json({ message: 'Word was deleted successfully' });
     } catch (error) {
       next(error);
     }
